fix: restructure integration tests and add local test runners (#1343)
* fix: restructure integration tests and add local test runners

* fix: code review comments fix

* chore: move ansible config and remove old setup scripts

* chore: move duplicate functions to common [run-int-tests] [run-ui-tests]

* fix: revert inventory timers [run-int-tests] [run-ui-tests]

* fix: permission change [run-int-tests] [run-ui-tests]

* fix: update .env

* fix: add recreate [run-int-tests] [run-ui-tests]

* chore: remove newgrp [run-int-tests] [run-ui-tests]

---------
