+++ .github/workflows/ci-main.yaml
@@ -185,7 +185,7 @@ jobs:
           sudo apt-get install snmp -y
           sudo apt-get install python3-dev -y
       - name: run automatic_setup_microk8s.sh
-        run: integration_tests/automatic_setup_microk8s.sh integration
+        run: integration_tests/scripts/automatic_setup_microk8s.sh integration
       - name: run tests
         working-directory: integration_tests
         run: |
@@ -244,7 +244,7 @@ jobs:
           sudo systemctl restart docker
           docker --version
       - name: run automatic_setup_compose.sh
-        run: integration_tests/automatic_setup_compose.sh integration
+        run: integration_tests/scripts/automatic_setup_compose.sh integration
       - name: run tests
         working-directory: integration_tests
         run: |
