--- integration_tests/.env
@@ -1,95 +1,0 @@
-# Deployment configuration
-SC4SNMP_IMAGE=snmp-local
-SC4SNMP_TAG="latest"
-SCHEDULER_CONFIG_FILE_ABSOLUTE_PATH=
-TRAPS_CONFIG_FILE_ABSOLUTE_PATH=
-INVENTORY_FILE_ABSOLUTE_PATH=
-COREFILE_ABS_PATH=
-COREDNS_ADDRESS=172.28.0.255
-# Set when testing with IPv6; leave empty otherwise (must be within IPAM_SUBNET_IPv6, e.g. fd02::1)
-COREDNS_ADDRESS_IPv6=
-SC4SNMP_VERSION=latest
-IPv6_ENABLED=false
-
-# Dependencies images
-COREDNS_IMAGE=coredns/coredns
-COREDNS_TAG=1.11.1
-MIBSERVER_IMAGE=ghcr.io/pysnmp/mibs/container
-MIBSERVER_TAG=latest
-REDIS_IMAGE=docker.io/redis
-REDIS_TAG=8.2.2
-MONGO_IMAGE=docker.io/bitnamilegacy/mongodb
-MONGO_TAG=7.0.14-debian-12-r3
-
-# Splunk instance configuration
-SPLUNK_HEC_HOST=
-SPLUNK_HEC_PROTOCOL=https
-SPLUNK_HEC_PORT=8088
-SPLUNK_HEC_TOKEN=
-SPLUNK_HEC_INSECURESSL=true
-SPLUNK_SOURCETYPE_TRAPS=sc4snmp:traps
-SPLUNK_SOURCETYPE_POLLING_EVENTS=sc4snmp:event
-SPLUNK_SOURCETYPE_POLLING_METRICS=sc4snmp:metric
-SPLUNK_HEC_INDEX_EVENTS=netops
-SPLUNK_HEC_INDEX_METRICS=netmetrics
-SPLUNK_HEC_PATH=/services/collector
-SPLUNK_AGGREGATE_TRAPS_EVENTS=false
-IGNORE_EMPTY_VARBINDS=false
-
-# Workers configration
-WALK_RETRY_MAX_INTERVAL=180
-WALK_MAX_RETRIES=5
-METRICS_INDEXING_ENABLED=false
-POLL_BASE_PROFILES=true
-IGNORE_NOT_INCREASING_OIDS=
-WORKER_LOG_LEVEL=INFO
-UDP_CONNECTION_TIMEOUT=3
-MAX_OID_TO_PROCESS=70
-
-# Worker Poller
-WORKER_POLLER_CONCURRENCY=4
-PREFETCH_POLLER_COUNT=1
-WORKER_POLLER_REPLICAS=2
-WORKER_POLLER_CPU_LIMIT=1
-WORKER_POLLER_MEMORY_LIMIT=500M
-WORKER_POLLER_CPU_RESERVATIONS=0.5
-WORKER_POLLER_MEMORY_RESERVATIONS=250M
-
-# Worker Sender
-WORKER_SENDER_CONCURRENCY=4
-PREFETCH_SENDER_COUNT=30
-WORKER_SENDER_REPLICAS=1
-WORKER_SENDER_CPU_LIMIT=1
-WORKER_SENDER_MEMORY_LIMIT=500M
-WORKER_SENDER_CPU_RESERVATIONS=0.5
-WORKER_SENDER_MEMORY_RESERVATIONS=250M
-
-# Worker Trap
-WORKER_TRAP_CONCURRENCY=4
-PREFETCH_TRAP_COUNT=30
-RESOLVE_TRAP_ADDRESS=false
-MAX_DNS_CACHE_SIZE_TRAPS=500
-TTL_DNS_CACHE_TRAPS=1800
-WORKER_TRAP_REPLICAS=2
-WORKER_TRAP_CPU_LIMIT=1
-WORKER_TRAP_MEMORY_LIMIT=500M
-WORKER_TRAP_CPU_RESERVATIONS=0.5
-WORKER_TRAP_MEMORY_RESERVATIONS=250M
-
-# Inventory configuration
-INVENTORY_LOG_LEVEL=INFO
-CHAIN_OF_TASKS_EXPIRY_TIME=500
-ENABLE_FULL_WALK=true
-
-# Traps configuration
-SNMP_V3_SECURITY_ENGINE_ID=80003a8c04
-TRAPS_PORT=162
-TRAP_LOG_LEVEL=INFO
-
-# Scheduler configuration
-SCHEDULER_LOG_LEVEL=INFO
-
-#Secrets
-SECRET_FOLDER_PATH=
-ENABLE_TRAPS_SECRETS=false
-ENABLE_WORKER_POLLER_SECRETS=false
