+++ integration_tests/README.md
@@ -30,3 +30,7 @@ ssh -i .ssh/snmp-ssh-key.pem ubuntu@54.90.167.146
 ```
 
 You can observe the progress in `~/splunk-connect-for-snmp/integration_tests/pytest.log`
+
+### Run simulator with SNMPv3
+
+If you want to walk SNMP V3 device place "test" value as secret in inventory.csv, copy sample_v3_values from integration tests into working directory, rename to secrets and edit 7 filed under secrets/snmpv3/test directory
