--- integration_tests/README_ENV_SETUP
@@ -1,80 +1,0 @@
-Deployment script on Nova
-
-
-#!/bin/bash
-
-deploy_image(){
-    git clone https://github.com/splunk/$1
-    cd $1
-    git checkout $2
-    git pull
-    sudo poetry install
-    sudo poetry build
-    sudo docker build . -t $3
-    cd ..
-    sudo docker save $3 > myimage.tar
-    sudo microk8s ctr image import myimage.tar
-}
-
-sudo pip3 install poetry
-
-deploy_image splunk-connect-for-snmp-poller feat/async_rework my-poller
-deploy_image splunk-connect-for-snmp-traps async_rework my-traps
-
-sed -i 's+ghcr.io/splunk/splunk-connect-for-snmp-poller:1+my-poller:latest+g' splunk-connect-for-snmp/deploy/sc4snmp/internal/scheduler-deployment.yaml
-sed -i 's+ghcr.io/splunk/splunk-connect-for-snmp-poller:1+my-poller:latest+g' splunk-connect-for-snmp/deploy/sc4snmp/internal/worker-deployment.yaml
-sed -i 's+ghcr.io/splunk/splunk-connect-for-snmp-traps:1+my-traps:latest+g' splunk-connect-for-snmp/deploy/sc4snmp/internal/traps-deployment.yaml
-
-sed -i 's/Always/Never/g' splunk-connect-for-snmp/deploy/sc4snmp/internal/scheduler-deployment.yaml
-sed -i 's/Always/Never/g' splunk-connect-for-snmp/deploy/sc4snmp/internal/worker-deployment.yaml
-sed -i 's/Always/Never/g' splunk-connect-for-snmp/deploy/sc4snmp/internal/traps-deployment.yaml
-
-sudo microk8s.kubectl delete -n sc4snmp -f splunk-connect-for-snmp/deploy/sc4snmp/internal/scheduler-deployment.yaml
-sudo microk8s.kubectl delete -n sc4snmp -f splunk-connect-for-snmp/deploy/sc4snmp/internal/worker-deployment.yaml
-
-sudo microk8s.kubectl apply -n sc4snmp -f splunk-connect-for-snmp/deploy/sc4snmp/internal/scheduler-deployment.yaml
-sudo microk8s.kubectl apply -n sc4snmp -f splunk-connect-for-snmp/deploy/sc4snmp/internal/worker-deployment.yaml
-
-&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&
-
-How to run Poller on local machine
-
-docker run -d -p 5672:5672 rabbitmq
-docker run -d -p 27017:27017 -v ~/data:/data/db mongo
-
-docker run -e MONGO_URI=mongodb://localhost:27017 -e MIBS_FILES_URL=http://host.docker.internal:5000/files/asn1/@mib@ -p 5000:5000  ghcr.io/splunk/splunk-connect-for-snmp-mib-server:develop
-
-docker run  -p 8881:8881 -p 8882:8882 -e SPLUNK_HEC_URL=https://10.202.11.190:8088/services/collector  -e SPLUNK_HEC_TOKEN=00000000-0000-0000-0000-000000000000 -e SPLUNK_HEC_TLS_SKIP_VERIFY=true -v "$(pwd)"/test:/config quay.io/signalfx/splunk-otel-collector:0.21.0 --config=/config/otel-collector-config.yaml
-
-docker run -p 161:161/udp tandrup/snmpsim
-
-
-To start celery worker:
-export MIBS_SERVER_URL=http://localhost:5000
-export MIBS_FILES_URL=http://localhost:5000/files/asn1/@mib@
-export OTEL_SERVER_LOGS_URL=http://localhost:8881
-export OTEL_SERVER_METRICS_URL=http://localhost:8882
-
-
-
-export MONGO_URI=mongodb://localhost:27017
-export CELERY_BROKER_URL=amqp://guest:guest@localhost:5672//
-export OTEL_SERVER_LOGS_URL=http://localhost:8881
-export OTEL_SERVER_METRICS_URL=http://localhost:8882
-export SPLUNK_HEC_URI=https://localhost:8088/services/collector
-export SPLUNK_HEC_INDEX_METRICS=em_metrics
-export SPLUNK_HEC_INDEX_EVENTS=em_events
-export SPLUNK_HEC_TOKEN=796bb611-0189-47dc-a5f8-25fe7be42c00
-
-celery -A splunk_connect_for_snmp_poller.manager.celery_client worker -l INFO -n worker1
-
-
-poetry run celery -A splunk_connect_for_snmp.poller beat -S celerybeatmongo.schedulers.MongoScheduler --loglevel=DEBUG
-poetry run celery -A splunk_connect_for_snmp.poller worker --loglevel=DEBUG
-
-To start scheduler:
-Parameters
--c ../config.yaml -i ../inventory.csv --event_index em_events --metric_index em_metrics --meta_index em_meta
-
-Environment variables
-PYTHONUNBUFFERED=1;CELERY_BROKER_URL=localhost;MONGO_URI=mongodb://localhost:27017
