+++ integration_tests/README_INTEGRATION_TESTS
@@ -15,7 +15,7 @@ Right now we support only NOVA.
 - sudo ./install_microk8s.sh
 - su <user_name>
 - cd ~/splunk-connect-for-snmp/integration_tests
-- ./automatic_setup.sh
+- ./automatic_setup_microk8s.sh
 
 Run with terraform and ansible. Use local_run.sh script. It has some prerequisites - AWS env variables and
 ansible, terraform commands. It will return error and information if you are missing something. You should
