+++ integration_tests/ansible_config/ansible.cfg
