+++ integration_tests/ansible_config/inventory.tmpl
