+++ integration_tests/ansible_config/main.tf.tmpl
