+++ integration_tests/scripts/prepare_splunk.sh
